// packages/soporte-it/src/equipos/dto/create-equipo.dto.d.ts
import { EstadoEquipo } from '../../entities/equipo.entity';
export declare class CreateEquipoDto {
    numeroActivo?: number;
    aCargoDe?: string;
    sector?: string;
    hostname?: string;
    windowsUserId?: string;
    fabricante?: string;
    modelo?: string;
    ramGb?: string;
    sistemaOperativo?: string;
    procesador?: string;
    firmwareUefi?: string;
    graficos?: string;
    almacenamiento?: string;
    adaptadorRed?: string;
    ipv6?: string;
    controladorUsbHost?: string;
    estado?: EstadoEquipo;
    notas?: string;
    usuarioPlatId?: string;
}
//# sourceMappingURL=create-equipo.dto.d.ts.map